Better number parsing for size property
git-svn-id: http://svn.osgeo.org/geotools/trunk@5248 e5c1c795-43da-0310-a71f-fac65c449510

+++ geotools2/geotools-src/renderer/src/org/geotools/renderer/style/SLDStyleFactory.java
@@ -369,7 +369,7 @@ public class SLDStyleFactory {
         //int size = ((Number) sldGraphic.getSize().getValue(feature)).intValue();
         int size;
         try{
-            size = Integer.parseInt(sldGraphic.getSize().getValue(feature).toString());
+            size = (int) Double.parseDouble(sldGraphic.getSize().getValue(feature).toString());
         }
         catch(NumberFormatException nfe){
             size = 1;
